+++ R/airsis_loadAnnual.R
@@ -7,6 +7,7 @@
 #' @param archiveBaseUrl Base URL for monitoring v2 data files.
 #' @param archiveBaseDir Local base directory for monitoring v2 data files.
 #' @param QC_negativeValues Type of QC to apply to negative values.
+#' @param QC_removeSuspectData Removes monitors determined to be misbehaving.
 #'
 #' @return A \emph{mts_monitor} object with AIRSIS data. (A list with
 #' \code{meta} and \code{data} dataframes.)
@@ -23,6 +24,16 @@
 #' For the most recent data in the last 10 days, use \code{airsis_loadLatest()}.
 #'
 #' For daily updates covering the most recent 45 days, use \code{airsis_loadDaily()}.
+#'
+#' @note
+#' With \code{QC_removeSuspectData = TRUE} (the default), data is checked for
+#' monitors with values of 2000 ug/m3. Some older AIRSIS timeseries contain only
+#' values of 0, 1000, 2000, 3000, 4000 and 5000. Data from these deployments
+#' passed instrument-level QC checks but these timeseries generally do not
+#' represent valid data and should be removed.
+#'
+#' Only those personally familiar with the individual instrument deployments
+#' should work with the "suspect" data.
 #'
 #' @seealso \code{\link{airsis_loadDaily}}
 #' @seealso \code{\link{airsis_loadLatest}}
@@ -53,7 +64,8 @@ airsis_loadAnnual <- function(
     "monitoring/v2"
   ),
   archiveBaseDir = NULL,
-  QC_negativeValues = c("zero", "na", "ignore")
+  QC_negativeValues = c("zero", "na", "ignore"),
+  QC_removeSuspectData = TRUE
 ) {
 
   parameterName <- "PM2.5"
@@ -149,20 +161,31 @@ airsis_loadAnnual <- function(
 
   # ----- Apply QC -------------------------------------------------------------
 
+  # Handle negative values
   if ( QC_negativeValues == "zero" ) {
-
     monitor <- monitor_replaceValues(monitor, data < 0, 0)
-
   } else if ( QC_negativeValues == "na" ) {
-
     monitor <- monitor_replaceValues(monitor, data < 0, as.numeric(NA))
+  }
+
+  # NOTE:  Several monitors in 2015 have values only at 0, 1000, 2000, ..., 5000
+  if ( QC_removeSuspectData ) {
+
+    badIDs <-
+      monitor %>%
+      monitor_selectWhere( function(x) { any(x == 2000, na.rm = TRUE) } ) %>%
+      monitor_getMeta() %>%
+      dplyr::pull(.data$deviceDeploymentID)
+
+    goodIDs <- setdiff(monitor$meta$deviceDeploymentID, badIDs)
 
   }
+
+  monitor <- monitor %>% monitor_select(goodIDs)
 
   # ----- Return ---------------------------------------------------------------
 
   return(monitor)
-
 
 }
 
@@ -171,10 +194,11 @@ airsis_loadAnnual <- function(
 if ( FALSE ) {
 
 
-  year <- 2021
+  year <- 2015
   archiveBaseUrl <- "https://airfire-data-exports.s3.us-west-2.amazonaws.com/monitoring/v2"
   archiveBaseDir <- NULL
   QC_negativeValues = "zero"
+  QC_removeSuspectData = TRUE