+++ R/airsis_loadLatest.R
@@ -6,6 +6,7 @@
 #' @param archiveBaseUrl Base URL for monitoring v2 data files.
 #' @param archiveBaseDir Local base directory for monitoring v2 data files.
 #' @param QC_negativeValues Type of QC to apply to negative values.
+#' @param QC_removeSuspectData Removes monitors determined to be misbehaving.
 #'
 #' @return A \emph{mts_monitor} object with AIRSIS data. (A list with
 #' \code{meta} and \code{data} dataframes.)
@@ -23,6 +24,16 @@
 #' For daily updates covering the most recent 45 days, use \code{airsis_loadDaily()}.
 #'
 #' For data extended more than 45 days into the past, use \code{airsis_loadAnnual()}.
+#'
+#' @note
+#' With \code{QC_removeSuspectData = TRUE} (the default), data is checked for
+#' monitors with values of 3000 ug/m3. Some older AIRSIS timeseries contain only
+#' values of 0, 1000, 2000, 3000, 4000 and 5000. Data from these deployments
+#' passed instrument-level QC checks but these timeseries generally do not
+#' represent valid data and should be removed.
+#'
+#' Only those personally familiar with the individual instrument deployments
+#' should work with the "suspect" data.
 #'
 #' @seealso \code{\link{airsis_loadAnnual}}
 #' @seealso \code{\link{airsis_loadDaily}}
@@ -47,7 +58,8 @@ airsis_loadLatest <- function(
     "monitoring/v2"
   ),
   archiveBaseDir = NULL,
-  QC_negativeValues = c("zero", "na", "ignore")
+  QC_negativeValues = c("zero", "na", "ignore"),
+  QC_removeSuspectData = TRUE
 ) {
 
   parameterName <- "PM2.5"
@@ -139,20 +151,31 @@ airsis_loadLatest <- function(
 
   # ----- Apply QC -------------------------------------------------------------
 
+  # Handle negative values
   if ( QC_negativeValues == "zero" ) {
-
     monitor <- monitor_replaceValues(monitor, data < 0, 0)
-
   } else if ( QC_negativeValues == "na" ) {
-
     monitor <- monitor_replaceValues(monitor, data < 0, as.numeric(NA))
+  }
+
+  # NOTE:  Several monitors in 2015 have values only at 0, 1000, 2000, ..., 5000
+  if ( QC_removeSuspectData ) {
+
+    badIDs <-
+      monitor %>%
+      monitor_selectWhere( function(x) { any(x == 3000, na.rm = TRUE) } ) %>%
+      monitor_getMeta() %>%
+      dplyr::pull(.data$deviceDeploymentID)
+
+    goodIDs <- setdiff(monitor$meta$deviceDeploymentID, badIDs)
 
   }
+
+  monitor <- monitor %>% monitor_select(goodIDs)
 
   # ----- Return ---------------------------------------------------------------
 
   return(monitor)
-
 
 }